commiting so i can pull from main. Right now only have eft and right icons
