+++ src/components/BottomDrawer.tsx
@@ -12,6 +12,7 @@ import {
 import SearchBars from "./SearchBars";
 import { useCoords } from "../data/CoordsContext";
 import { BottomDrawerStyle } from "../styles/BottomDrawerStyle";
+import DirectionsSteps from "./DirectionsSteps";
 import { SearchBar } from "react-native-elements";
 
 const { height, width } = Dimensions.get("window");
@@ -28,39 +29,6 @@ function BottomDrawer({
   children: ReactNode;
   drawerHeight: Animated.Value;
 }) {
-  const { routeData: routeCoordinates,isTransit,setIsTransit} = useCoords();
-  const [htmlInstructions, setHtmlInstructions] = useState<string[]>([]);
-
-  useEffect(() => {
-
-
-    if (routeCoordinates && routeCoordinates.length > 0) {
-      const instructions = routeCoordinates[0].legs[0].steps.map((step: any) => { return step.html_instructions.replace(/<.*?>/g, '').replace(/(?=Destination)/gi, '. '); });
-     
-      //Declaring the transit route, they are more precise 
-      let detailedInstructions: string[]=[];
-      const number = routeCoordinates[0].legs[0].steps.length;
-      const instructionsGeneral = routeCoordinates[0].legs[0].steps.map((step: any) => { return step.html_instructions.replace(/<.*?>/g, ''); });
-      for (let i=0; i<number; i++){
-        const detailedHtmlInstructions = routeCoordinates[0]?.legs[0]?.steps[i]?.steps?.map((step: any) => {
-          return step.html_instructions.replace(/<.*?>/g, '').replace(/(?=Destination)/gi, '. '); });
-          if (detailedHtmlInstructions==undefined){
-              detailedInstructions= detailedInstructions.concat(instructionsGeneral[i]);
-          }else {
-          detailedInstructions=detailedInstructions.concat(instructionsGeneral[i]);
-          detailedInstructions=detailedInstructions.concat(detailedHtmlInstructions);}
-      }
-    //Set htmlInstruction to the transit one if chosen mode is transit
-    if (isTransit==true){
-      setHtmlInstructions(detailedInstructions);
-    } else {
-      setHtmlInstructions(instructions);
-    }
-    console.log(htmlInstructions);
-    } else {
-      setHtmlInstructions([]);
-    }
-  }, [routeCoordinates]);
 
   const isExpanded = useRef<boolean>(true);
 
@@ -108,16 +76,7 @@ function BottomDrawer({
       <View {...panResponder.panHandlers} style={BottomDrawerStyle.dragHandle}>
         <View style={BottomDrawerStyle.dragIndicator} />
         <SearchBars />
-        <View style={BottomDrawerStyle.listContent}>
-        <ScrollView>
-        {htmlInstructions.length > 0 &&
-          htmlInstructions.map((instruction, index) => (
-            <Text key={index} style={BottomDrawerStyle.instructionsList}>
-              {instruction}
-            </Text>
-          ))}
-           </ScrollView>
-          </View>
+       <DirectionsSteps />
       </View>
       <View style={BottomDrawerStyle.contentContainer}>{children}</View>
     </Animated.View>
